changed folder structure and seprated the constants
