+++ constants/colors/Colors.ts
@@ -1,0 +1,4 @@
+const COLORS = {
+  primary: '#522446',
+};
+export default COLORS;
