+++ constants/strings/Strings.ts
@@ -1,0 +1,4 @@
+const STRINGS = {
+  continue: 'Continue',
+};
+export default STRINGS;
