+++ screens/Auth/auth_labels.ts
@@ -1,0 +1,5 @@
+const AUTH_LABELS = {
+  title: 'Verify your mobile number',
+  subtitle: `Enter your number so that we know you're you`,
+};
+export default AUTH_LABELS;
